+++ lib/concurrent-edge.rb
@@ -10,3 +10,4 @@ require 'concurrent/edge/atomic_markable_reference'
 require 'concurrent/edge/lock_free_linked_set'
 
 require 'concurrent/edge/promises'
+require 'concurrent/edge/cancellation'
